Added .gitignore to avoid git reporting untracked build files/results/input etc Added LIM search scripts Added .gitkeep to divimp/cases so the directory will be created when the repo is cloned Added -fallow-argument-mismatch flag to various Makefile.gcc compile in current gfortran with warnings and not errors (each case needs to be looked at sometime since some may be bugs) Added support utilities in support/efit_wall to pull wall definitions from G efits (based on code found at D3D - author unknown) and a python script to convert a two column data file to a DIVIMP wall description. Added README.md with information on cloning OEDGE as well as information on git and git commands for those new to using git (like me :) ). Modified - bug fix in mod_fp_data where code was reading/writing an integer array using a routine expecting real Modified - bug fix in mod_out_unstruc - real*8 was being read using a routine expecting real - changed variable to real Added - code to analyse Carre/Sonnet grids to obtain the grid "magic numbers" by finding the mesh breaks at pfz/core knots and pfz/sol, core/sol rings. Added - Carre/Sonnet grid read routine uses field separators to parse element input - eg "(",",",")" - these changes were required by Carre/Sonnet grid incompatibilities in the ecrim5.F DIVIMP grid output for Carre module in SOLPS-ITER. DIVIMP was adjusted to work with format changes and no longer requires the "GEOM:" input at the beginning of the grid file. Modified - SOL22 no longer reports imaginary values encountered in the solver as an error as long as they are resolved by reducing the step size. Added - a CINOUT routine was added to utility_com.f to support reading/writing of character arrays. This was needed because hc_global_routines was using a routine for real numbers to read/write character arrays (this needs more testing - but should be low impact so pushing the change) Added mod_readgrid.o to the DIVIMP OBJECT_LIST Modified div.f - replaced a few co-terminated labelled do loops with the do/end do construct from modern fortran. Eventually all of these should be replaced. Modified tau.f - added the new grid reading code - removed some of the "goto" constructs - replacing them with appropriate IF statement termination. Added some checks for wall size > maxpts,mves and checking to see if psifl has appropriate values on both sides of the separatrix to ensure it has been read correctly. Replaced some calls to RZERO with <var>=0.0 - this should be done throughout the code since RZERO is no longer needed. Modified - added some progress indicators to the run_testcases script to echo when each has finished.

+++ out6/src/outa00.f
@@ -2761,7 +2761,7 @@ c
       real*8 :: totcore_poloidal_area(maxnks),
      >          totcore_radial_area(maxnrs)
       real*8 :: omp_reg,xpt_reg
-      !real*8 :: psi1_reg,psi2_reg
+      !real  :: psi1_reg,psi2_reg
       integer :: in
 
       integer :: nr
